+++ include/arch/x86/device_info/PIC.cpp
@@ -108,7 +108,7 @@ void IRQ_clear_mask(u8 IRQline)
 void PIC::enableIRQ(u8 irq_id)
 {
 
-    LOG("PIC: Enabling IRQ", static_cast<u16>(irq_id));
+    LOG("PIC: Enabling IRQ", irq_id);
 
     if (irq_id < 8)
     {
@@ -121,8 +121,8 @@ void PIC::enableIRQ(u8 irq_id)
         mask2 = mask2 & ~(1 << (irq_id - 8));
         outb(PIC2_DATA, mask2);
     }
-    LOG("\tmask1: ", static_cast<u16>(mask1), " mask2: ", static_cast<u16>(mask2));
-    LOG("PIC: IRQ", static_cast<u16>(irq_id), " enabled");
+    LOG("\tmask1: ", mask1, " mask2: ", mask2);
+    LOG("PIC: IRQ", irq_id, " enabled");
 }
 
 
@@ -141,7 +141,7 @@ void PIC::disableIRQ(const u8 irq_id)
         mask2 = mask2 | (1 << (irq_id-8));
         outb(PIC2_DATA, mask2);
     }
-    LOG("mask1: ", static_cast<u16>(mask1), " mask2: ", static_cast<u16>(mask2));
+    LOG("mask1: ", mask1, " mask2: ", mask2);
 }
 
 void PIC::enableAll()
